LPS-115637 simplify questions screens and add disabled search
