+++ modules/apps/questions/questions-web/src/main/resources/META-INF/resources/js/hooks/useSection.es.js
@@ -21,7 +21,7 @@ export default function useSection(sectionTitle, siteKey) {
 
 	useEffect(() => {
 		getSections(sectionTitle, siteKey).then(setSection);
-	}, [sectionTitle, siteKey])
+	}, [sectionTitle, siteKey]);
 
 	return section;
 }
